fix compile issues for tests and use serial_test crate

+++ swap/src/bitcoin/wallet.rs
@@ -1004,9 +1004,10 @@ mod tests {
 
     #[test]
     fn given_depth_0_should_meet_confirmation_target_one() {
+
         let script = ScriptStatus::Confirmed(Confirmed { depth: 0 });
 
-        let confirmed = script.is_confirmed_with(1);
+        let confirmed = script.is_confirmed_with(1 as u32);
 
         assert!(confirmed)
     }
@@ -1015,7 +1016,7 @@ mod tests {
     fn given_confirmations_1_should_meet_confirmation_target_one() {
         let script = ScriptStatus::from_confirmations(1);
 
-        let confirmed = script.is_confirmed_with(1);
+        let confirmed = script.is_confirmed_with(1 as u32);
 
         assert!(confirmed)
     }
@@ -1034,7 +1035,7 @@ mod tests {
     fn given_depth_0_should_return_0_blocks_left_until_1() {
         let script = ScriptStatus::Confirmed(Confirmed { depth: 0 });
 
-        let blocks_left = script.blocks_left_until(1);
+        let blocks_left = script.blocks_left_until(1 as u32);
 
         assert_eq!(blocks_left, 0)
     }
@@ -1043,7 +1044,7 @@ mod tests {
     fn given_depth_1_should_return_0_blocks_left_until_1() {
         let script = ScriptStatus::Confirmed(Confirmed { depth: 1 });
 
-        let blocks_left = script.blocks_left_until(1);
+        let blocks_left = script.blocks_left_until(1 as u32);
 
         assert_eq!(blocks_left, 0)
     }
@@ -1052,7 +1053,7 @@ mod tests {
     fn given_depth_0_should_return_1_blocks_left_until_2() {
         let script = ScriptStatus::Confirmed(Confirmed { depth: 0 });
 
-        let blocks_left = script.blocks_left_until(2);
+        let blocks_left = script.blocks_left_until(2 as u32);
 
         assert_eq!(blocks_left, 1)
     }
